requirements.txt
streamlit
tensorflow
pillow
numpy 
opencv-python
scikit-learn
matplotlib
pandas
huggingface-hub
kagglehub
seaborn